+++ pydy/codegen/matrix_generator.py
@@ -1,15 +1,10 @@
 #!/usr/bin/env python
 
-"""This module contains source code dedicated to generating C code from
-matrices generated from sympy.physics.mechanics."""
-
-import os
 import itertools
 
 import sympy as sm
 import sympy.physics.mechanics as me
 from sympy.printing.codeprinter import CodePrinter
-from sympy.printing.octave import OctaveCodePrinter
 
 from ..utils import wrap_and_indent, find_dynamicsymbols
 
@@ -198,56 +193,3 @@ class MatrixGenerator(object):
                 outputs += '\n\n'  # space between each output
 
         self.code_blocks['outputs'] = outputs
-
-
-class OctaveMatrixGenerator(MatrixGenerator):
-    """This class generates Octave/Matlab source files that simultaneously
-    numerically evaluate any number of SymPy matrices.
-
-    """
-
-    _idx_start = 1
-    _idx_delim = "()"
-    _base_printer = OctaveCodePrinter
-    _type_declar = ''  # prepended to variable introductions
-    _line_contin = ' ...'
-    _comment_char = '%'
-
-    # TODO : The first two lines will not wrap. For many inputs/outputs it
-    # would be nice to have some wrapping.
-    _m_template = """\
-function [{output_args}] = {prefix}({input_args})
-% function [{output_args}] = {prefix}({input_args})
-%
-{docstring}
-
-{subexprs}
-
-{outputs}
-
-end
-"""
-
-    def doprint(self, prefix='eval_mats'):
-        """Returns a string that implements the function.
-
-        Parameters
-        ==========
-        prefix : string, optional
-            The name of the Octave/Matlab function.
-
-        """
-        self.code_blocks['prefix'] = prefix
-
-        return self._m_template.format(**self.code_blocks)
-
-    def write(self, prefix='eval_mats', path=None):
-        """Writes the <prefix>.m file to disc at the give path location."""
-
-        if path is None:
-            path = os.getcwd()
-
-        text = self.doprint(prefix=prefix)
-
-        with open(os.path.join(path, prefix + '.m'), 'w') as f:
-            f.write(text)
